fix(tri_inv): drop tri_inverse_impl
